+++ UnityUtils/Scripts/Hotkeys/Editor/CompileProject.cs
@@ -1,0 +1,9 @@
+using UnityEditor;
+using UnityEditor.Compilation;
+
+public static class CompileProject {
+    [MenuItem("File/Compile _F2")]
+    static void Compile() {
+        CompilationPipeline.RequestScriptCompilation(RequestScriptCompilationOptions.CleanBuildCache);
+    }
+}